+++ api/weatherData/weatherData.go
@@ -3,17 +3,18 @@ package weatherData
 import (
 	"encoding/json"
 	"errors"
+	"fmt"
 	"main/db"
 	"main/debug"
 	"net/http"
 	"strings"
-	"time"
 )
 
 // WeatherData structure stores current and predicted weather data for a day.
 //
 // Functionality: Handler, get
 type WeatherData struct {
+	Updated string `json:"updated"`
 	Now struct {
 		AirTemperature      float64 `json:"air_temperature"`
 		CloudAreaFraction   float64 `json:"cloud_area_fraction"`
@@ -65,9 +66,23 @@ func (weatherData *WeatherData) Handler(w http.ResponseWriter, r *http.Request) 
 		debug.ErrorMessage.Print(w)
 		return
 	}
-	//check if data was in database and either read data or get new data
-	if exist {
+	//get status on timeframe and branch if an error occurred
+	withinTimeframe, err := db.CheckDate(data.Time, 6)
+	//check if data is in database and if it's usable then either read data or get new data
+	if exist && withinTimeframe {
+		fmt.Println("Got data from database")
+		if err != nil {
+			debug.ErrorMessage.Update(
+				http.StatusInternalServerError, 
+				"WeatherData.Handler() -> Database.CheckDate() -> parsing time",
+				err.Error(),
+				"Unknown",
+			)
+			debug.ErrorMessage.Print(w)
+			return
+		}
 		err = weatherData.readData(data.Container)
+		weatherData.Updated = data.Time
 		if err != nil {
 			debug.ErrorMessage.Update(
 				http.StatusInternalServerError, 
@@ -93,9 +108,8 @@ func (weatherData *WeatherData) Handler(w http.ResponseWriter, r *http.Request) 
 		}
 		//send data to database
 		var data db.Data
-		data.Time = time.Now().String()
 		data.Container = weatherData
-		err = db.DB.Add("WeatherData", id, data)
+		date, err := db.DB.Add("WeatherData", id, data)
 		if err != nil {
 			debug.ErrorMessage.Update(
 				http.StatusInternalServerError, 
@@ -106,6 +120,7 @@ func (weatherData *WeatherData) Handler(w http.ResponseWriter, r *http.Request) 
 			debug.ErrorMessage.Print(w)
 			return
 		}
+		weatherData.Updated = date
 	}
 	//update header to JSON and set HTTP code
 	w.Header().Set("Content-Type", "application/json")