+++ ChangeLog.md
@@ -73,3 +73,15 @@
 [0.21.0]    **development**:    WeatherData Handler sends data to database
 
 [0.22.0]    **development**:    Added database reader
+
+[0.23.0]    **development**:    Added time validation on data stored in database
+
+[0.24.0]    **development**:    Refactored HandlerCoords to check firestore for existing data.
+
+[0.25.0]    **development**:    Fixed time validation when errors occur
+
+[0.26.0]    **development**:    Added updated field in WeatherData
+
+[0.27.0]    **development**:    Added ability to add and get holidays from the database
+
+[0.28.0]    **development**:    Saving information from restCountries localy instead of firebase